% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Qpress_files.R
\name{make.qnm}
\alias{make.qnm}
\title{Signed diagraphs to Qpress}
\usage{
make.qnm(modmat)
}
\arguments{
\item{modmat}{Output from signed diagraphs}
}
\value{
Output object for Qpress
}
\description{
Uses the output from a signed diagraph to set up the input for Qpress.
}
\seealso{
Other QNM_comparisons: \code{\link{MM2Qpress}}
}
