% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Qpress_files.R
\name{MM2Qpress}
\alias{MM2Qpress}
\title{Mental Modeler to Qpress}
\usage{
MM2Qpress(MMout)
}
\arguments{
\item{MMout}{Output from Mental modeler}
}
\value{
Output object for Qpress
}
\description{
Uses the output from Mental Modeler to set up the input for Qpress.
}
\seealso{
Other QNM_comparisons: \code{\link{make.qnm}}
}
